partial refactor with code migrated to correspondence_controller_util
